requirements.txt
allure-pytest
pytest~=8.2.2
selene>=2.0.0rc2
